-Internationalization in customer -refactoring

+++ RestaurantApi/src/main/java/com/yp/service/CategoryService.java
@@ -5,19 +5,24 @@ import com.yp.dto.CategoryDto;
 import com.yp.dto.ProductDto;
 import com.yp.entity.Category;
 import com.yp.entity.Product;
+import com.yp.exception.BusinessRuleException;
+import com.yp.exception.ContentNotFoundException;
 import com.yp.mapper.CategoryMapper;
 import com.yp.mapper.MediaMapper;
 import com.yp.mapper.ProductMapper;
 import com.yp.repos.CategoryRepository;
 import com.yp.repos.ProductRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.cache.annotation.CacheEvict;
+import org.springframework.cache.annotation.CachePut;
+import org.springframework.cache.annotation.Cacheable;
+import org.springframework.cache.annotation.EnableCaching;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Slice;
 import org.springframework.stereotype.Service;
 
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.List;
+import java.util.*;
+
 
 @Service
 public class CategoryService {
@@ -36,53 +41,68 @@ public class CategoryService {
     @Autowired
     private MediaMapper mediaMapper;
 
-    public List<CategoryDto> getAllCategory(){
+    @Cacheable(value = "CategoryCache")
+    public List<CategoryDto> getAllCategory(String lang){
         List<Category> categories =  categoryRepository.findAll();
-        List<CategoryDto> categoryDtos = new ArrayList<>();
-        categories.forEach(category -> {
-            CategoryDto categoryDto = categoryMapper.toDto(category);
-            categoryDtos.add(categoryDto);
-        });
-        return categoryDtos;
+        return categoryMapper.toDtoList(categories);
     }
 
+    @Cacheable(value = "CategoryCache",key = "'CATEGORY_CACHE_BY_ID'.concat(#id)")
     public CategoryDto getCategory(Long id){
-        Category category = categoryRepository.findById(id).get();
-        CategoryDto categoryDto = categoryMapper.toDto(category);
-        return categoryDto;
+        if(id==null) throw new BusinessRuleException("No Id was sent");
+        Optional<Category> optionalCategory = categoryRepository.findById(id);
+        if(optionalCategory.isEmpty()) throw new ContentNotFoundException("Category not found");
+        return categoryMapper.toDto(optionalCategory.get());
+
     }
 
+    @CacheEvict(value = "CategoryCache", allEntries = true)
     public Category addCategory(CategoryDto categoryDto){
+        if(categoryDto==null) throw new BusinessRuleException("Category cannot be null");
         Category cat = categoryMapper.toEntity(categoryDto);
         return categoryRepository.save(cat);
     }
 
+    //@CacheEvict(value = "CategoryCache", key = "'CATEGORY_CACHE_BY_ID'.concat(#id)")
+    @CacheEvict(value = "CategoryCache", allEntries = true)
     public Category editCategory(Long id, CategoryDto categoryDto){
-        Category cat = categoryRepository.findById(id).get();
-        cat.setName(categoryDto.getName());
-        cat.setMedia(mediaMapper.toEntity(categoryDto.getMedia()));
+        Optional<Category> optionalCategory = categoryRepository.findById(id);
+        if(optionalCategory.isEmpty()) {
+            throw new ContentNotFoundException("Category not found");
+        }
+        Category cat = optionalCategory.get();
+        if(!cat.getName().equals(categoryDto.getName())){
+            cat.setName(categoryDto.getName());
+        }
+        if(cat.getMedia()!=mediaMapper.toEntity(categoryDto.getMedia())){
+            cat.setMedia(mediaMapper.toEntity(categoryDto.getMedia()));
+        }
         return categoryRepository.save(cat);
     }
 
+    @CacheEvict(value = "CategoryCache", allEntries = true)
     public void deleteCategory(Long id){
+        if(id==null) throw new BusinessRuleException("No id was given");
         categoryRepository.deleteById(id);
     }
 
-    public List<ProductDto> getProductsWithId(Long id){
-        Category category = categoryRepository.findById(id).get();
-        List<Product> products = productRepository.findAllByCategoriesIn(Arrays.asList(category));
-        List<ProductDto> productDtos = new ArrayList<>();
-        products.forEach(product -> {
-            ProductDto productDto = productMapper.toDto(product);
-            productDtos.add(productDto);
-        });
-        return productDtos;
+    public List<ProductDto> getProductsWithCategoryId(Long id){
+        Optional<Category> optionalCategory = categoryRepository.findById(id);
+        if(optionalCategory.isEmpty()) {
+            return null;
+        }
+        Category category = optionalCategory.get();
+        List<Product> products = productRepository.findAllByCategoriesIn(Collections.singletonList(category));
+        return productMapper.toDtoList(products);
     }
 
-    public Slice<ProductDto> getProductSliceWithId(Long id, Pageable pageable){
-        Category category = categoryRepository.findById(id).get();
-        Slice<ProductDto> productSlice = productRepository.findAllByCategories(category, pageable).map(productMapper::toDto);
-        return productSlice;
+    public Slice<ProductDto> getProductsSliceWithCategoryId(Long id, Pageable pageable){
+        Optional<Category> optionalCategory = categoryRepository.findById(id);
+        if(optionalCategory.isEmpty()){
+            return null;
+        }
+        Category category = optionalCategory.get();
+        return productRepository.findAllByCategories(category, pageable).map(productMapper::toDto);
     }
 
 }
